+++ src/components/header/logo.tsx
@@ -1,6 +1,7 @@
 import { SITE_NAME } from '@/consts'
 import { CENTERED_ROW_CLS, FOCUS_RING_CLS, HEADER_ICON_SIZE_CLS } from '@/theme'
 import { cn } from '@lib/shadcn-utils'
+import { BaseImage } from '../base-image'
 import { BaseLink } from '../link/base-link'
 
 const LOGO_CLS = cn(FOCUS_RING_CLS, HEADER_ICON_SIZE_CLS, CENTERED_ROW_CLS)
@@ -12,7 +13,7 @@ export function Logo() {
   return (
     <BaseLink href="/" className={LOGO_CLS} aria-label="Home">
       <div className="flex h-6 w-6 flex-row items-center justify-center rounded-full bg-theme fill-white">
-        <img
+        <BaseImage
           src="/favicon.svg"
           width={512}
           height={512}